Adds a new method of finding climbable Surfaces
Instead of using a sphere trace and and then doing another line trace to
determine the angle. I've decided to draw eight rays with their origin
around the hip of the player character. This accomplishes the task of
finding the angle of the slope of potential climbable surfaces quite
well.